Added auth pages layout
